pyproject.toml
[tool.poetry]
name = "phonebook-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.13"
phonenumbers = "^8.13.39"
django-phonenumber-field = "^7.3.0"
django-bootstrap-v5 = "^1.0.11"
faker = "^25.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
